template (caseview edit)
