--- edit/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-class EditConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'edit'
