+++ review/views.py
@@ -1,6 +1,6 @@
 from django.http import JsonResponse
 from django.shortcuts import render
-from edit.models import Changes
+from caseview.models import Changes
 
 def review_page(request):    
     return render(request, "review.html")
